Change whitelist argument for addresses to size_t.
Previously, the pixel addresses for the whitelist were given in Python
by a flex.int() array.  This is now changed to flex.size_t() on the
ground that it is the more correct type for array addresses.  Address
lists can now be computed in Python with

bool_array.iselection()

rather than the more awkward

bool_array.iselection().as_int().

Corresponding changes are made in C++, CUDA and Kokkos code.

+++ simtbx/gpu/tst_unified.py
@@ -147,7 +147,7 @@ class several_wavelength_case:
             ichannels = flex.int(range(len(self.flux))),
             gpu_amplitudes = gpu_channels_singleton,
             gpu_detector = gpu_detector,
-            pixel_active_list_ints = flex.int(range(NN)),
+            pixel_active_list_ints = flex.size_t(range(NN)),
             weights = self.frac/len(self.frac)
       )
     per_image_scale_factor = self.domains_per_crystal # 1.0
